pyproject.toml
[project]
name = "socialconnect"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pyautogui>=0.9.54",
    "pyperclip>=1.8.2",
    "python-dotenv>=1.0.0",
    "pywhatkit>=5.4",
]
